+++ rust/datafusion-wasm/src/lib.rs
@@ -4,9 +4,6 @@ use arrow::array::RecordBatch;
 use arrow::datatypes::Schema;
 use arrow::ipc::reader::StreamReader;
 use arrow::ipc::writer::StreamWriter;
-use datafusion::execution::SessionStateBuilder;
-use datafusion::physical_optimizer::PhysicalOptimizerRule;
-use datafusion::physical_optimizer::optimizer::PhysicalOptimizer;
 use datafusion::prelude::*;
 use micromegas_tracing::prelude::*;
 use wasm_bindgen::prelude::*;
@@ -42,28 +39,8 @@ impl WasmQueryEngine {
         ensure_tracing();
         info!("WasmQueryEngine created");
 
-        // Work around a DataFusion 52.1 bug where the LimitPushdown physical
-        // optimizer rule removes GlobalLimitExec without actually pushing the
-        // fetch into DataSourceExec, causing LIMIT to be silently ignored.
-        // Fixed upstream in https://github.com/apache/datafusion/pull/20048
-        // but not yet released.
-        // TODO: remove after upgrading DataFusion past 52.1
-        // https://github.com/madesroches/micromegas/issues/809
-        let filtered_rules = PhysicalOptimizer::default()
-            .rules
-            .into_iter()
-            .filter(|rule: &Arc<dyn PhysicalOptimizerRule + Send + Sync>| {
-                rule.name() != "LimitPushdown"
-            })
-            .collect::<Vec<_>>();
-
-        let config = SessionConfig::default().with_information_schema(true);
-        let state = SessionStateBuilder::new()
-            .with_config(config)
-            .with_default_features()
-            .with_physical_optimizer_rules(filtered_rules)
-            .build();
-        let ctx = SessionContext::new_with_state(state);
+        let ctx =
+            SessionContext::new_with_config(SessionConfig::default().with_information_schema(true));
         micromegas_datafusion_extensions::register_extension_udfs(&ctx);
         Self { ctx }
     }